+++ tests/Service/NetworkHelperTest.php
@@ -17,9 +17,9 @@ class NetworkHelperTest extends TestCase
    */
   public function testLocationInfo() {
     $helper = new NetworkHelper();
-    $result = $helper->getLocationInfoByIp('178.197.235.71'); // IP is in Zürich...
+    $result = $helper->getLocationInfoByIp('178.197.235.71'); // IP is in Bern...
     $this->assertEquals('CH', $result['country']);
-    $this->assertEquals('Zurich', $result['city']);
+    $this->assertEquals('Bern', $result['city']);
   }
 
   /**